Add custom macros for the memory-functions - calloc, - realloc, - free, and - strdup
These functions will not directly improve memory handling but they will log if something will go wrong *before* it actually goes wrong. This will ease debugging when user provide log snippets especially in ENOMEM (no memory available) situations.
