+++ parser.c
@@ -1282,96 +1282,6 @@ int findClientID(const char *client)
 	return clientID;
 }
 
-void validate_access(const char * name, int pos, bool testmagic, int line, const char * function, const char * file)
-{
-	int limit = 0;
-	if(name[0] == 'c') limit = counters.clients_MAX;
-	else if(name[0] == 'd') limit = counters.domains_MAX;
-	else if(name[0] == 'q') limit = counters.queries_MAX;
-	else if(name[0] == 'o') limit = counters.overTime_MAX;
-	else if(name[0] == 'f') limit = counters.forwarded_MAX;
-	else if(name[0] == 'w') limit = counters.wildcarddomains;
-	else { logg("Validator error (range)"); killed = 1; }
-
-	if(pos >= limit || pos < 0)
-	{
-		logg("FATAL ERROR: Trying to access %s[%i], but maximum is %i", name, pos, limit);
-		logg("             found in %s() (line %i) in %s", function, line, file);
-	}
-	// Don't test magic byte if detected potential out-of-bounds error
-	else if(testmagic)
-	{
-		unsigned char magic = 0x00;
-		if(name[0] == 'c') magic = clients[pos].magic;
-		else if(name[0] == 'd') magic = domains[pos].magic;
-		else if(name[0] == 'q') magic = queries[pos].magic;
-		else if(name[0] == 'o') magic = overTime[pos].magic;
-		else if(name[0] == 'f') magic = forwarded[pos].magic;
-		else { logg("Validator error (magic byte)"); killed = 1; }
-		if(magic != MAGICBYTE)
-		{
-			logg("FATAL ERROR: Trying to access %s[%i], but magic byte is %x", name, pos, magic);
-			logg("             found in %s() (line %i) in %s", function, line, file);
-		}
-	}
-}
-
-void validate_access_oTfd(int timeidx, int forwardID, int line, const char * function, const char * file)
-{
-	// Determine if there is enough space for saving the current
-	// forwardID in the overTime data structure, allocate space otherwise
-	validate_access("overTime", timeidx, true, __LINE__, __FUNCTION__, __FILE__);
-	if(overTime[timeidx].forwardnum <= forwardID)
-	{
-		// Reallocate more space for forwarddata
-		overTime[timeidx].forwarddata = realloc(overTime[timeidx].forwarddata, (forwardID+1)*sizeof(*overTime[timeidx].forwarddata));
-		// Initialize new data fields with zeroes
-		int j;
-		for(j = overTime[timeidx].forwardnum; j <= forwardID; j++)
-		{
-			overTime[timeidx].forwarddata[j] = 0;
-			memory.forwarddata++;
-		}
-		// Update counter
-		overTime[timeidx].forwardnum = forwardID + 1;
-	}
-
-	int limit = overTime[timeidx].forwardnum;
-	if(forwardID >= limit || forwardID < 0)
-	{
-		logg("!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!");
-		logg("FATAL ERROR: Trying to access overTime.forwarddata[%i], but maximum is %i", forwardID, limit);
-		logg("             found in %s() (line %i) in %s", function, line, file);
-	}
-}
-
-void validate_access_oTcl(int timeidx, int clientID, int line, const char * function, const char * file)
-{
-	// Determine if there is enough space for saving the current
-	// clientID in the overTime data structure, allocate space otherwise
-	if(overTime[timeidx].clientnum <= clientID)
-	{
-		// Reallocate more space for clientdata
-		overTime[timeidx].clientdata = realloc(overTime[timeidx].clientdata, (clientID+1)*sizeof(*overTime[timeidx].clientdata));
-		// Initialize new data fields with zeroes
-		int i;
-		for(i = overTime[timeidx].clientnum; i <= clientID; i++)
-		{
-			overTime[timeidx].clientdata[i] = 0;
-			memory.clientdata++;
-		}
-		// Update counter
-		overTime[timeidx].clientnum = clientID + 1;
-	}
-	int limit = overTime[timeidx].clientnum;
-	if(clientID >= limit || clientID < 0)
-	{
-		logg("!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!");
-		logg("FATAL ERROR: Trying to access overTime.clientdata[%i], but maximum is %i", clientID, limit);
-		logg("             found in %s() (line %i) in %s", function, line, file);
-	}
-}
-
 void reresolveHostnames(void)
 {
 	int clientID;
